Extract common concurrent test utils
